refactor(core): Reorganize code, apply design patterns, and parallelize tests
This commit includes several changes:

- `signature.go` has been moved from `internal/core/consensus` to `internal/core/block`.
- The `VerifyBlockSignature` function has been renamed to `VerifyBlock` for consistency with `VerifyTransaction`.
- The code in `transaction.go` has been split into three files: `add_file.go` for `AddFileTransaction`, `delete_file.go` for `DeleteFileTransaction`, and `transaction.go` for common transaction elements.
- The Factory design pattern has been applied to unify `Deserialize[struct name]` functions behind a single `DeserializeTransaction` function.
- The Template Method design pattern has been applied to unify `VerifyTransaction` functions behind a single `VerifyTransaction` function for all transactions.
- `t.Parallel()` has been added to AES key tests to enable parallel execution of tests.
