+++ internal/core/block/signature.go
@@ -1,20 +1,19 @@
-package consensus
+package block
 
 import (
 	"log"
 
-	"github.com/pierreleocadie/SecuraChain/internal/core/block"
 	"github.com/pierreleocadie/SecuraChain/pkg/ecdsa"
 )
 
-// VerifyBlockSignature checks if the block signature is valid
-func VerifyBlockSignature(currentBlock *block.Block) bool {
+// VerifyBlock checks if the block signature is valid
+func VerifyBlock(currentBlock *Block) bool {
 	if len(currentBlock.Header.Signature) == 0 {
 		log.Printf("Block validation failed: Signature is empty")
 		return false
 	}
 
-	headerHash := block.ComputeHash(currentBlock)
+	headerHash := ComputeHash(currentBlock)
 
 	ecdsaPublicKey, err := ecdsa.PublicKeyFromBytes(currentBlock.MinerAddr)
 	if err != nil {
