+++ internal/core/transaction/client_announcement.go
@@ -9,6 +9,13 @@ import (
 	"github.com/pierreleocadie/SecuraChain/pkg/ecdsa"
 )
 
+// ClientAnnouncementFactory implements the TransactionFactory interface
+type ClientAnnouncementFactory struct{}
+
+func (f *ClientAnnouncementFactory) CreateTransaction(data []byte) (Transaction, error) {
+	return DeserializeClientAnnouncement(data)
+}
+
 type ClientAnnouncement struct {
 	AnnouncementId        uuid.UUID `json:"announcementId"`        // Announcement ID - UUID
 	OwnerAddress          []byte    `json:"ownerAddress"`          // Owner address - ECDSA public key
@@ -18,6 +25,21 @@ type ClientAnnouncement struct {
 	Checksum              []byte    `json:"checksum"`              // Checksum - SHA256
 	OwnerSignature        []byte    `json:"ownerSignature"`        // Owner signature - ECDSA signature
 	AnnouncementTimestamp int64     `json:"announcementTimestamp"` // Announcement timestamp - Unix timestamp
+	TransactionVerifier             // embed TransactionVerifier struct to inherit VerifyTransaction method
+}
+
+func (a *ClientAnnouncement) Serialize() ([]byte, error) {
+	return json.Marshal(a)
+}
+
+// Override SpecificData for ClientAnnouncement
+func (a *ClientAnnouncement) SpecificData() ([]byte, error) {
+	// Remove signature and serialize
+	signature := a.OwnerSignature
+	a.OwnerSignature = []byte{}
+	defer func() { a.OwnerSignature = signature }() // Restore after serialization
+
+	return json.Marshal(a)
 }
 
 func NewClientAnnouncement(keyPair ecdsa.KeyPair, filename []byte, extension []byte, fileSize uint64, checksum []byte) *ClientAnnouncement {
@@ -49,10 +71,6 @@ func NewClientAnnouncement(keyPair ecdsa.KeyPair, filename []byte, extension []b
 	return announcement
 }
 
-func (a *ClientAnnouncement) Serialize() ([]byte, error) {
-	return json.Marshal(a)
-}
-
 func DeserializeClientAnnouncement(data []byte) (*ClientAnnouncement, error) {
 	var announcement ClientAnnouncement
 	err := json.Unmarshal(data, &announcement)
@@ -60,26 +78,4 @@ func DeserializeClientAnnouncement(data []byte) (*ClientAnnouncement, error) {
 		return nil, err
 	}
 	return &announcement, nil
-}
-
-func VerifyClientAnnouncement(announcement *ClientAnnouncement) bool {
-	// Remove signature from announcement before verifying
-	ownerSignature := announcement.OwnerSignature
-	announcement.OwnerSignature = []byte{}
-
-	announcementBytes, err := json.Marshal(announcement)
-	if err != nil {
-		return false
-	}
-	announcementHash := sha256.Sum256(announcementBytes)
-
-	// Restore signature
-	announcement.OwnerSignature = ownerSignature
-
-	ownerAddress, err := ecdsa.PublicKeyFromBytes(announcement.OwnerAddress)
-	if err != nil {
-		return false
-	}
-
-	return ecdsa.VerifySignature(ownerAddress, announcementHash[:], ownerSignature)
 }
