+++ internal/core/transaction/delete_file.go
@@ -1,0 +1,77 @@
+package transaction
+
+import (
+	"crypto/sha256"
+	"encoding/json"
+	"time"
+
+	"github.com/google/uuid"
+	"github.com/pierreleocadie/SecuraChain/pkg/ecdsa"
+)
+
+// DeleteFileTransactionFactory implements the TransactionFactory interface
+type DeleteFileTransactionFactory struct{}
+
+func (f *DeleteFileTransactionFactory) CreateTransaction(data []byte) (Transaction, error) {
+	return DeserializeDeleteFileTransaction(data)
+}
+
+type DeleteFileTransaction struct {
+	TransactionId        uuid.UUID `json:"transactionId"`        // Transaction ID - UUID
+	OwnerAddress         []byte    `json:"ownerAddress"`         // Owner address - ECDSA public key
+	FileCID              []byte    `json:"fileCID"`              // File CID - SHA256
+	TransactionSignature []byte    `json:"transactionSignature"` // Transaction signature - ECDSA signature
+	TransactionTimestamp int64     `json:"transactionTimestamp"` // Transaction timestamp - Unix timestamp
+	TransactionVerifier            // embed TransactionVerifier struct to inherit VerifyTransaction method
+}
+
+func (t *DeleteFileTransaction) Serialize() ([]byte, error) {
+	return json.Marshal(t)
+}
+
+// Override SpecificData for DeleteFileTransaction
+func (t *DeleteFileTransaction) SpecificData() ([]byte, error) {
+	// Remove signature from transaction before verifying
+	signature := t.TransactionSignature
+	t.TransactionSignature = []byte{}
+
+	defer func() { t.TransactionSignature = signature }() // Restore after serialization
+
+	return json.Marshal(t)
+}
+
+func NewDeleteFileTransaction(keyPair ecdsa.KeyPair, fileCID []byte) *DeleteFileTransaction {
+	ownerAddressBytes, err := keyPair.PublicKeyToBytes()
+	if err != nil {
+		return nil
+	}
+
+	transaction := &DeleteFileTransaction{
+		TransactionId:        uuid.New(),
+		OwnerAddress:         ownerAddressBytes,
+		FileCID:              fileCID,
+		TransactionTimestamp: time.Now().Unix(),
+	}
+
+	transactionBytes, err := json.Marshal(transaction)
+	if err != nil {
+		return nil
+	}
+	transactionHash := sha256.Sum256(transactionBytes)
+
+	transaction.TransactionSignature, err = keyPair.Sign(transactionHash[:])
+	if err != nil {
+		return nil
+	}
+
+	return transaction
+}
+
+func DeserializeDeleteFileTransaction(data []byte) (*DeleteFileTransaction, error) {
+	var transaction DeleteFileTransaction
+	err := json.Unmarshal(data, &transaction)
+	if err != nil {
+		return nil, err
+	}
+	return &transaction, nil
+}
